+++ install.sh
@@ -88,7 +88,7 @@ install_debian_deps() {
                        libmp3lame-dev libvorbis-dev libopus-dev libvpx-dev \
                        libavahi-client-dev libavahi-common-dev libasound2-dev \
                        libfreetype-dev libharfbuzz-dev libfontconfig1-dev libfribidi-dev \
-                       intel-gpu-tools xvfb x11vnc x11-xserver-utils xfonts-base \
+                       intel-gpu-tools xvfb x11vnc x11-xserver-utils xfonts-base xdotool \
                        libnss3 libnspr4 libatk1.0-0 libatk-bridge2.0-0 libcups2 libdrm2 \
                        libxkbcommon0 libxcomposite1 libxdamage1 libxfixes3 libxrandr2 \
                        libgbm1 libpango-1.0-0 libcairo2 libasound2
@@ -102,7 +102,7 @@ install_rhel_deps() {
                    x264-devel x265-devel openssl-devel libva-devel libdrm-devel \
                    lame-devel libvorbis-devel opus-devel libvpx-devel \
                    avahi-devel alsa-lib-devel freetype-devel harfbuzz-devel fontconfig-devel fribidi-devel \
-                   intel-gpu-tools xorg-x11-server-Xvfb x11vnc xorg-x11-server-utils \
+                   intel-gpu-tools xorg-x11-server-Xvfb x11vnc xorg-x11-server-utils xdotool \
                    nss nspr atk at-spi2-atk cups-libs libdrm libxkbcommon \
                    libXcomposite libXdamage libXfixes libXrandr mesa-libgbm pango cairo alsa-lib
 }
@@ -114,7 +114,7 @@ install_arch_deps() {
                                  x264 x265 openssl libva libdrm \
                                  lame libvorbis opus libvpx \
                                  avahi alsa-lib freetype2 harfbuzz fontconfig fribidi \
-                                 intel-gpu-tools xorg-server-xvfb x11vnc xorg-xset python nodejs npm \
+                                 intel-gpu-tools xorg-server-xvfb x11vnc xorg-xset python nodejs npm xdotool \
                                  nss nspr atk at-spi2-atk cups libdrm libxkbcommon \
                                  libxcomposite libxdamage libxfixes libxrandr mesa pango cairo alsa-lib
 }